update to esm and fix bugs in functions resolver and add endpoint to list all exist functions.
