--- example-functions/example.js
@@ -1,6 +1,0 @@
-exports.hello = {
-    path: '/hello',
-    onRequest: (request, response) => {
-        response.json({message: "Hello"});
-    }
-}
