+++ gulpfile.js
@@ -13,7 +13,7 @@ function startDev(cb) {
             LOGS: '1',
             URL_TAR: 'https://github.com/fahamutech/bfast-database/releases/download/latest/bfast-database-latest.tgz',
             NPM_TAR: 'bfast-database',
-            MONGO_URL: 'mongodb://localhost/daas',
+            DATABASE_URI: 'mongodb://localhost/daas',
             PORT: '3004',
             PRODUCTION: '1',
             PROJECT_ID: 'faas',
