+++ src/function/_example.cjs
@@ -1,4 +1,4 @@
-exports.helloCommonjs = {
+exports.hello_commonjs__ = {
     path: '/example-functions/hello-c',
     onRequest: (request, response) => {
         response.json({message: 'Hello, World! --> commonjs'});
