+++ src/function/_example.js
@@ -1,4 +1,4 @@
-exports.helloNormal = {
+exports.hello_normal__ = {
     path: '/example-functions/hello',
     onRequest: (request, response) => {
         response.json({message: 'Hello, World!'});
