+++ src/function/_health.mjs
@@ -1,4 +1,4 @@
-export const functionsHealth = {
+export const functions__health__ = {
     path: '/functions-health',
     onRequest: (request, response) => {
         response.json({message: 'running'});
